requirements.txt
amqp==2.3.2
billiard==3.5.0.5
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-cleanup==2.1.0
django-cool-pagination==0.3.0
django-phone-field==1.6.0
django-sslserver==0.20
django-widget-tweaks==1.4.3
idna==2.7
mysqlclient==1.3.14
oauthlib==2.1.0
pathlib==1.0.1
Pillow==5.3.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
Unipath==1.1
urllib3==1.24.1
vine==1.1.4
